+++ MvvmCross.iOS.Support/Views/Expandable/AccordionSectionExpandableController.cs
@@ -1,7 +1,6 @@
-using System;
 using System.Collections.Generic;
 
-namespace MvvmCross.iOS.Support.Views.ExpandableInternal
+namespace MvvmCross.iOS.Support.Views.Expandable
 {
 	class AccordionSectionExpandableController : SectionExpandableController
 	{
